feat: Add image training pipeline, ONNX export, improved dashboard, and reorganized docs
- Added image model training script and config (configs/image/efficientnet_b3.yaml)
- Created lightweight dataset generator (< 10 GB) for quick testing
- Exported trained video model to ONNX format (43.8 MB) for production deployment
- Enhanced dashboard UI/UX: removed blank spaces, added system capabilities, live inference
- Improved real-time detector with float32 fixes
- Reorganized documentation into docs/ folder for cleaner structure
- Fixed ensemble config with proper ONNX paths
- Updated inference pipeline with real model integration
- Added batch processing capabilities to dashboard
- Model: EfficientNet-B3 trained on Celeb-DF dataset (Val Loss: 0.0616)

+++ src/realtime/detector.py
@@ -221,7 +221,7 @@ class RealtimeDetector:
                     
                     # Softmax
                     probs = np.exp(logits) / np.sum(np.exp(logits))
-                    predictions['video'] = float(probs[1])  # Fake probability
+                    predictions['video'] = float(probs[1])  # Fake probability (class 1 = FAKE)
                     
                 except Exception as e:
                     logger.error(f"Video inference error: {e}")
@@ -274,7 +274,7 @@ class RealtimeDetector:
         transposed = np.transpose(normalized, (2, 0, 1))
         
         # Add batch dimension
-        tensor = np.expand_dims(transposed, axis=0)
+        tensor = np.expand_dims(transposed, axis=0).astype(np.float32)
         
         return tensor
